fix: move projects to separate file and change imports
