+++ features/bookmarks/DialogAddProductsToFolder.tsx
@@ -25,7 +25,7 @@ import {
 // Import actual data
 import { departments } from "@/app/(app)/departments/data";
 import { sampleProducts } from "@/app/(app)/products/data";
-import { projects } from "@/app/(app)/projects/page";
+import { projects } from "@/app/(app)/projects/data";
 import {
   PiStackPlusDuotone,
   PiCirclesThreePlusDuotone,
