+++ features/products/CreateProductDialog.tsx
@@ -26,7 +26,7 @@ import {
 
 // Import actual data from the pages
 import { departments } from "@/app/(app)/departments/data";
-import { projects } from "@/app/(app)/projects/page";
+import { projects } from "@/app/(app)/projects/data";
 
 export default function CreateProductDialog() {
   const id = useId();
