+++ features/products/product/product-information/ProductInformationEditProductDialog.tsx
@@ -20,7 +20,7 @@ import { Plus, X } from "lucide-react";
 
 // Import actual data from the pages
 import { departments } from "@/app/(app)/departments/data";
-import { projects } from "@/app/(app)/projects/page";
+import { projects } from "@/app/(app)/projects/data";
 
 interface ProductData {
   productId: string;
